[refactor] 경로 수정
